sony: common: Move omni custom configuration to common
Move common omni configuration to common repo then avoid
unnecessary quantity of the same change on devices repos.

DRY.

Remove radio.mk since it is not a common stuff
because we have Wi-Fi only devices (like karin_windy).
radio.mk is common only for GSM/LTE devices
and it is not aligned with common stuff proposal.
Also we have only one property in radio.mk
it does not justify a file creation so
move the property to GSM/LTE devices and still
using only one line.

Signed-off-by: Humberto Borba <humberos@gmail.com>
Change-Id: I3bf3871145ef6c6a6ff5bfb2bb3225a05b351013